rename gemini/utils gemini/gemini_utils to follow openai module
